fix(account): unblock Vercel prod deploy — Next.js 16 cacheComponents + dynamic({ ssr: false }) (BRO-1229)
broomva.tech Vercel production deploys have been FAILING for ~22h
(since BRO-1213 passkey UI merge at `a058535`). Multiple feature PRs
blocked from going live, including BRO-1224 (just-merged transparent
Tier-1 refresh) — the new `/api/auth/lifegw-token` endpoint isn't
live in production because Vercel can't ship it.

## Two stacked Next.js 16 incompatibilities

### 1. `nextDynamic({ ssr: false })` in Server Components

`apps/broomva/app/account/security/passkey/page.tsx:27` (server-side
async page) was calling `nextDynamic(..., { ssr: false, loading: ... })`.
Next.js 16 dropped support for `ssr: false` in Server Component
contexts — the option is now Client-Component-only.

**Fix**: extracted the dynamic-import + skeleton into a new client-
component file `apps/broomva/components/account/passkey-card-lazy.tsx`
(`"use client"`). Page imports `PasskeyCardLazy` as a regular
component. Behaviour preserved: same lazy chunk (WebAuthn ceremony
stays out of the shared client shell), same skeleton during chunk
load, same props passthrough. Only the file that owns the
`nextDynamic` call moved.

### 2. `export const dynamic = "force-dynamic"` + cacheComponents

`apps/broomva/next.config.ts` has `cacheComponents: true`. Next.js 16
treats `export const dynamic = "force-dynamic"` as incompatible with
that flag — `Route segment config "dynamic" is not compatible with
nextConfig.cacheComponents. Please remove it.` Three files were
affected:

- `apps/broomva/app/account/layout.tsx:15`
- `apps/broomva/app/account/page.tsx:24`
- `apps/broomva/app/account/security/passkey/page.tsx:24`

**Fix**: removed the explicit `export const dynamic` from all three.
Each page already calls `await headers()`, which automatically opts
into dynamic rendering under cacheComponents — the explicit
declaration was redundant *and* now blocks the build. Replaced with
a comment explaining the rationale.

## Verification

* `next build` locally past the account/* error site. Remaining build
  failures (`Error: cookies.secret must be at least 32 chars`,
  `Failed to collect page data for /api/files/upload`) are unrelated
  env-config issues that exist only when local lacks Vercel's secrets
  — Vercel CI has them set.
* `tsc --noEmit -p tsconfig.json` clean for the changed paths.

## Acceptance

* Vercel production deploy of `main` succeeds after this merges
* `/account/security/passkey` still renders with skeleton fallback
* `https://broomva.tech/api/auth/lifegw-token` returns 405 (BRO-1224
  endpoint live)

Closes BRO-1229.
